requirements.txt
Flask==3.1.0
Flask-SocketIO==5.5.1
eventlet==0.39.1
mss==10.0.0
Pillow==11.1.0
pyautogui==0.9.54
cryptography==44.0.0
